refactor: fixture db 패키지 추가 및 도메인 추가
