+++ src/test/java/roomescape/fixture/db/ReservationDateFixture.java
@@ -1,4 +1,4 @@
-package roomescape.fixture;
+package roomescape.fixture.db;
 
 import java.time.LocalDate;
 import roomescape.reservation.domain.ReservationDate;
