+++ src/test/java/roomescape/time/service/ReservationTimeServiceTest.java
@@ -13,10 +13,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import roomescape.common.CleanUp;
-import roomescape.fixture.MemberDbFixture;
-import roomescape.fixture.ReservationDateTimeDbFixture;
-import roomescape.fixture.ReservationTimeDbFixture;
-import roomescape.fixture.ThemeDbFixture;
+import roomescape.fixture.db.MemberDbFixture;
+import roomescape.fixture.db.ReservationDateTimeDbFixture;
+import roomescape.fixture.db.ReservationTimeDbFixture;
+import roomescape.fixture.db.ThemeDbFixture;
 import roomescape.global.exception.InvalidArgumentException;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.Reservation;
